+++ samples/segment/get_subscribers.php
@@ -13,7 +13,7 @@ $result = $wrap->get_subscribers('Added since', 1, 50, 'email', 'asc', true);
 //$result = $wrap->get_subscribers(date('Y-m-d', strtotime('-30 days')), 
 //  page number, page size, order by, order description);
 
-echo "Result of GET /api/v3.1/segments/{segment id}/active\n<br />";
+echo "Result of GET /api/v3.3/segments/{segment id}/active\n<br />";
 if($result->was_successful()) {
     echo "Got subscribers\n<br /><pre>";
     var_dump($result->response);